+++ frontend/src/modules/help/router.js
@@ -1,0 +1,7 @@
+const Help = () => import('./')
+
+export default {
+  path: '/hjaelp',
+  name: 'Help',
+  component: Help
+}
