+++ frontend/src/modules/query/router.js
@@ -1,0 +1,7 @@
+const QueryList = () => import('./')
+
+export default {
+  path: '/forespoergsler',
+  name: 'QueryList',
+  component: QueryList
+}
