+++ frontend/src/modules/router.js
@@ -1,0 +1,11 @@
+import helpRouter from './help/router'
+import organisationMapperRouter from './organisationMapper/router'
+import queryRouter from './query/router'
+import timeMachineRouter from './timeMachine/router'
+
+export default [
+  helpRouter,
+  organisationMapperRouter,
+  queryRouter,
+  timeMachineRouter
+]
