+++ frontend/src/views/employee/EmployeeDetail.vue
