+++ frontend/src/views/employee/EmployeeDetailTabs.vue
